move schema generator to new app package to make it reusable in other modules
